refactor: combine slack dir w/ service dir
